access_control/app/requirements/prod.txt
# Everything needed in production

# Flask
alembic==0.9.9
aniso8601==3.0.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-HTTPAuth==3.2.3
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.8
psycopg2==2.7.5
Werkzeug==0.14.1
gunicorn==19.8.1

# Utilities
requests==2.18.4
validate-email==1.3
phonenumbers==8.9.9


# Testing
pytest==3.6.1
WebTest==2.0.29
factory-boy==2.11.1
